requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.1
pytz==2024.1
sqlparse==0.4.4
